client_serving/requirements.txt
fastapi
uvicorn
pillow
pyyaml
python-multipart